major: Move processors to separate module
