+++ gradle.properties
@@ -1,4 +1,4 @@
-projectVersion=4.6.5-SNAPSHOT
+projectVersion=5.0.0-SNAPSHOT
 projectGroup=io.micronaut.security
 
 micronautDocsVersion=2.0.0
