--- security-annotations/src/main/resources/META-INF/services/io.micronaut.inject.annotation.AnnotationMapper
@@ -1,3 +1,0 @@
-io.micronaut.security.annotation.JakartaPermitAllAnnotationMapper
-io.micronaut.security.annotation.JakartaRolesAllowedAnnotationMapper
-io.micronaut.security.annotation.JakartaDenyAllAnnotationMapper
