+++ security-processor/build.gradle.kts
@@ -1,0 +1,14 @@
+plugins {
+    id("io.micronaut.build.internal.security-module")
+}
+
+dependencies {
+    compileOnly(mn.micronaut.core.processor)
+    implementation(projects.micronautSecurityAnnotations)
+}
+
+micronautBuild {
+    binaryCompatibility {
+        enabled.set(false)
+    }
+}
