+++ security-processor/src/main/java/io/micronaut/security/processor/JakartaDenyAllAnnotationMapper.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2017-2023 original authors
+ * Copyright 2017-2024 original authors
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,39 +13,40 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.micronaut.security.annotation;
+package io.micronaut.security.processor;
 
 import io.micronaut.core.annotation.AnnotationValue;
 import io.micronaut.core.annotation.Internal;
 import io.micronaut.inject.annotation.NamedAnnotationMapper;
 import io.micronaut.inject.visitor.VisitorContext;
+import io.micronaut.security.annotation.Secured;
 
 import java.lang.annotation.Annotation;
 import java.util.ArrayList;
 import java.util.List;
 
 /**
- * Allows using the {@link jakarta.annotation.security.RolesAllowed} annotation in Micronaut.
+ * Allows using the {@link jakarta.annotation.security.DenyAll} annotation in Micronaut.
  *
  * @author Fredrik Hov
  */
 @Internal
-public class JakartaRolesAllowedAnnotationMapper implements NamedAnnotationMapper {
+public class JakartaDenyAllAnnotationMapper implements NamedAnnotationMapper {
+
     @Override
     public String getName() {
-        return "jakarta.annotation.security.RolesAllowed";
+        return "jakarta.annotation.security.DenyAll";
     }
 
     @Override
     public List<AnnotationValue<?>> map(AnnotationValue<Annotation> annotation, VisitorContext visitorContext) {
-        String[] values = annotation.get("value", String[].class).orElse(new String[0]);
-
         List<AnnotationValue<?>> annotationValues = new ArrayList<>(1);
         annotationValues.add(
             AnnotationValue.builder(Secured.class)
-                .values(values)
+                .value("denyAll()")
                 .build()
         );
         return annotationValues;
     }
+
 }
