+++ security-processor/src/main/java/io/micronaut/security/processor/JakartaRolesAllowedAnnotationMapper.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2017-2023 original authors
+ * Copyright 2017-2024 original authors
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,38 +13,41 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.micronaut.security.annotation;
+package io.micronaut.security.processor;
 
 import io.micronaut.core.annotation.AnnotationValue;
 import io.micronaut.core.annotation.Internal;
 import io.micronaut.inject.annotation.NamedAnnotationMapper;
 import io.micronaut.inject.visitor.VisitorContext;
+import io.micronaut.security.annotation.Secured;
 
 import java.lang.annotation.Annotation;
 import java.util.ArrayList;
 import java.util.List;
 
 /**
- * Allows using the {@link jakarta.annotation.security.DenyAll} annotation in Micronaut.
+ * Allows using the {@link jakarta.annotation.security.RolesAllowed} annotation in Micronaut.
  *
  * @author Fredrik Hov
  */
 @Internal
-public class JakartaDenyAllAnnotationMapper implements NamedAnnotationMapper {
+public class JakartaRolesAllowedAnnotationMapper implements NamedAnnotationMapper {
+
     @Override
     public String getName() {
-        return "jakarta.annotation.security.DenyAll";
+        return "jakarta.annotation.security.RolesAllowed";
     }
 
     @Override
     public List<AnnotationValue<?>> map(AnnotationValue<Annotation> annotation, VisitorContext visitorContext) {
+        String[] values = annotation.get("value", String[].class).orElse(new String[0]);
+
         List<AnnotationValue<?>> annotationValues = new ArrayList<>(1);
         annotationValues.add(
             AnnotationValue.builder(Secured.class)
-                .value("denyAll()")
+                .values(values)
                 .build()
         );
         return annotationValues;
     }
-
 }
