+++ security-processor/src/main/resources/META-INF/services/io.micronaut.inject.annotation.AnnotationMapper
@@ -1,0 +1,3 @@
+io.micronaut.security.processor.JakartaPermitAllAnnotationMapper
+io.micronaut.security.processor.JakartaRolesAllowedAnnotationMapper
+io.micronaut.security.processor.JakartaDenyAllAnnotationMapper
