+++ security/build.gradle.kts
@@ -4,9 +4,9 @@ plugins {
 
 dependencies {
     annotationProcessor(mn.micronaut.graal)
-
     annotationProcessor(mnSerde.micronaut.serde.processor)
     annotationProcessor(mnValidation.micronaut.validation.processor)
+
     api(mnSerde.micronaut.serde.api) {
         exclude(group = "io.micronaut", module = "micronaut-json-core")
     }
@@ -20,6 +20,10 @@ dependencies {
     compileOnly(mn.micronaut.http.server)
     compileOnly(mn.micronaut.management)
     compileOnly(mn.jackson.databind)
+
+    testAnnotationProcessor(projects.micronautSecurityProcessor)
+    testAnnotationProcessor(mn.micronaut.inject.java)
+
     testCompileOnly(mnData.micronaut.data.processor)
     testImplementation(mnSql.h2)
     testImplementation(mnSql.micronaut.jdbc.hikari)
@@ -28,8 +32,6 @@ dependencies {
     testImplementation(mnReactor.micronaut.reactor)
     testImplementation(mn.micronaut.management)
     testImplementation(mn.micronaut.http.client)
-    testAnnotationProcessor(projects.micronautSecurityAnnotations)
-    testAnnotationProcessor(mn.micronaut.inject.java)
     testImplementation(mn.micronaut.http.server.netty)
     testImplementation(projects.testSuiteUtils)
     testImplementation(mn.snakeyaml)
