+++ settings.gradle
@@ -16,6 +16,7 @@ include "security"
 include "security-aot"
 include "security-annotations"
 include "security-jwt"
+include "security-processor"
 include "security-session"
 include "security-ldap"
 include "security-oauth2"
